requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
arrow==1.2.3
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.10.0
bs4==0.0.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
chatgptpy==1.0.6
click==8.1.3
colorama==0.4.6
cookiecutter==1.7.3
cssselect2==0.7.0
fake-useragent==1.1.1
fastapi==0.79.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
humanize==4.4.0
hyperframe==6.0.1
idna==3.4
jieba==0.42.1
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.3
mypy-extensions==0.4.3
nb-cli==0.6.7
nonebot-adapter-onebot==2.1.5
nonebot-plugin-abstract==1.0.5
nonebot-plugin-apscheduler==0.1.4
nonebot-plugin-crazy-thursday==0.2.6
nonebot-plugin-remake==0.2.7
nonebot-plugin-status==0.6.0
nonebot2==2.0.0b5
OpenAIAuth==0.0.6
pathspec==0.10.2
Pillow==9.3.0
pinyin==0.4.0
platformdirs==2.6.0
poyo==0.5.0
prompt-toolkit==3.0.36
psutil==5.9.4
pydantic==1.9.2
pyfiglet==0.8.post1
pygtrie==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
revChatGPT==0.0.36.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.19.1
svglib==1.4.1
text-unidecode==1.3
tinycss2==1.2.1
tls-client==0.1.5
tomli==2.0.1
tomlkit==0.10.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
